My attempt on CodeUtilities

+++ src/main/java/io/github/codeutilities/cosmetics/CosmeticRenderer.java
@@ -13,7 +13,6 @@ import net.minecraft.client.render.entity.model.PlayerEntityModel;
 import net.minecraft.client.render.model.BakedModel;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.client.util.math.Vector3f;
-import net.minecraft.util.math.Quaternion;
 
 import java.io.IOException;
 
@@ -22,27 +21,29 @@ public class CosmeticRenderer extends FeatureRenderer<AbstractClientPlayerEntity
     public CosmeticRenderer(FeatureRendererContext<AbstractClientPlayerEntity, PlayerEntityModel<AbstractClientPlayerEntity>> context) {
         super(context);
     }
-    
+
     public void render(MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i, AbstractClientPlayerEntity abstractClientPlayerEntity, float f, float g, float h, float j, float k, float l) {
-        if(!abstractClientPlayerEntity.isInvisible()){
+        if (!abstractClientPlayerEntity.isInvisible()) {
             BakedModel model = null;
             JsonObject attributes = null;
             try {
                 model = CosmeticHandler.getModelFromUUID(abstractClientPlayerEntity.getUuid());
                 attributes = CosmeticHandler.getModelAttributesFromUUID(abstractClientPlayerEntity.getUuid());
-            } catch (IOException ignoreCuzIfPlayerDoesntHaveCosmeticItWillGiveFileNotFoundException) { }
-            if(model != null){
+            } catch (IOException ignoreCuzIfPlayerDoesntHaveCosmeticItWillGiveFileNotFoundException) {
+            }
+            if (model != null) {
                 Vector3f translation = model.getTransformation().head.translation;
                 Vector3f scale = model.getTransformation().head.scale;
                 ModelPart head = getContextModel().head;
                 Vector3f rotation = model.getTransformation().head.rotation;
-                float scalex = scale.getX()-0.333333333f;
-                float scaley = scale.getY()-0.333333333f;
-                float scalez = scale.getZ()-0.333333333f;
-                matrixStack.translate((translation.getX()*-1)*scalex, (translation.getY()*-1)*scaley, (translation.getZ()*-1)*scalez);
-                if(attributes.get("type").getAsString().equals("head")) rotate(matrixStack, 0f, abstractClientPlayerEntity.isInSneakingPose() ? translation.getY()*0.675f+0.27f : translation.getY()*0.675f, 0f, head.pitch, head.yaw, head.roll);
+                float scalex = scale.getX() - 0.333333333f;
+                float scaley = scale.getY() - 0.333333333f;
+                float scalez = scale.getZ() - 0.333333333f;
+                matrixStack.translate((translation.getX() * -1) * scalex, (translation.getY() * -1) * scaley, (translation.getZ() * -1) * scalez);
+                if (attributes.get("type").getAsString().equals("head"))
+                    rotate(matrixStack, 0f, abstractClientPlayerEntity.isInSneakingPose() ? translation.getY() * 0.675f + 0.27f : translation.getY() * 0.675f, 0f, head.pitch, head.yaw, head.roll);
                 matrixStack.scale(scalex, scaley, scalez);
-                if(abstractClientPlayerEntity.isInSneakingPose()) matrixStack.translate(0f, 0.4f, 0f);
+                if (abstractClientPlayerEntity.isInSneakingPose()) matrixStack.translate(0f, 0.4f, 0f);
                 matrixStack.translate(-0.5f, -0.75f, -0.5f);
 
                 MinecraftClient.getInstance().getBlockRenderManager().getModelRenderer().render(
@@ -59,7 +60,7 @@ public class CosmeticRenderer extends FeatureRenderer<AbstractClientPlayerEntity
     }
 
     public void rotate(MatrixStack matrix, float pivotX, float pivotY, float pivotZ, float pitch, float yaw, float roll) {
-        matrix.translate((double)(pivotX), (double)(pivotY), (double)(pivotZ));
+        matrix.translate(pivotX, pivotY, pivotZ);
         if (roll != 0.0F) {
             matrix.multiply(Vector3f.POSITIVE_Z.getRadialQuaternion(roll));
         }
@@ -71,7 +72,7 @@ public class CosmeticRenderer extends FeatureRenderer<AbstractClientPlayerEntity
         if (pitch != 0.0F) {
             matrix.multiply(Vector3f.POSITIVE_X.getRadialQuaternion(pitch));
         }
-        matrix.translate((double)(pivotX*-1), (double)(pivotY*-1), (double)(pivotZ*-1));
+        matrix.translate(pivotX * -1, pivotY * -1, pivotZ * -1);
     }
 
 }
